get cli working
