+++ pluggable-architecture/git-data-fetcher/README.md
@@ -1,3 +1,11 @@
 # `git` Data Fetcher
 
 This command line program will go fetch data from the various hosted `git` platforms.
+
+## Usage
+
+```console
+python cli.py --url https://github.com/busy-beaver-dev/busy-beaver
+
+python cli.py --url https://gitlab.com/pycqa/flake8
+```
